refactored authenticateJWT to it's own function
